@@ -1,16 +1,15 @@
 import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
-import 'package:meta/meta.dart';
 import 'dart:math';
 
 class MyViewPort extends RenderSliverFillViewport {
   int itemCount;
 
-  MyViewPort(
-      {@required RenderSliverBoxChildManager childManager,
-      double viewportFraction = 1.0,
-      this.itemCount})
-      : super(childManager: childManager, viewportFraction: viewportFraction);
+  MyViewPort({
+    required RenderSliverBoxChildManager childManager,
+    double viewportFraction = 1.0,
+    required this.itemCount,
+  }) : super(childManager: childManager, viewportFraction: viewportFraction);
 
   @override
   int getMaxChildIndexForScrollOffset(double scrollOffset, double itemExtent) {
@@ -35,13 +34,12 @@ class MyViewPort extends RenderSliverFillViewport {
 class MySliverFillViewport extends SliverMultiBoxAdaptorWidget {
   /// Creates a sliver whose box children that each fill the viewport.
   //
-  const MySliverFillViewport(
-      {Key key,
-      @required SliverChildDelegate delegate,
-      this.viewportFraction = 1.0,
-      this.itemCount})
-      : assert(viewportFraction != null),
-        assert(viewportFraction > 0.0),
+  const MySliverFillViewport({
+    Key? key,
+    required SliverChildDelegate delegate,
+    this.viewportFraction = 1.0,
+    required this.itemCount,
+  })   : assert(viewportFraction > 0.0),
         super(key: key, delegate: delegate);
 
   /// The fraction of the viewport that each child should fill in the main axis.
@@ -55,7 +53,9 @@ class MySliverFillViewport extends SliverMultiBoxAdaptorWidget {
 
   @override
   RenderSliverFillViewport createRenderObject(BuildContext context) {
-    final SliverMultiBoxAdaptorElement element = context;
+    final SliverMultiBoxAdaptorElement element = SliverMultiBoxAdaptorElement(
+        context.widget as SliverMultiBoxAdaptorWidget);
+
     return new MyViewPort(
         childManager: element,
         itemCount: itemCount,
@@ -78,11 +78,13 @@ const PageScrollPhysics _kPagePhysics = PageScrollPhysics();
 
 class Ext extends PageView {}
 
+// ignore: must_be_immutable
 class MyPageView extends StatelessWidget {
   final SliverChildListDelegate childrenDelegate;
 
-  MyPageView({List<Widget> children})
-      : childrenDelegate = new SliverChildListDelegate(children);
+  MyPageView({
+    required List<Widget> children,
+  }) : childrenDelegate = new SliverChildListDelegate(children);
 
   PageController controller = new PageController();
 
