@@ -31,7 +31,7 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, required this.title}) : super(key: key);
 
   final String title;
 
@@ -40,7 +40,7 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  IndexController _controller;
+  late IndexController _controller;
   List<String> _types = [
     "AccordionTransformer",
     "ThreeDTransformer",
@@ -50,8 +50,8 @@ class _MyHomePageState extends State<MyHomePage> {
     "DeepthPageTransformer"
   ];
 
-  String _type;
-  FixedExtentScrollController controller;
+  late String _type;
+  late FixedExtentScrollController controller;
 
   double _viewportFraction = 1.0;
 
@@ -92,27 +92,36 @@ class _MyHomePageState extends State<MyHomePage> {
         children: <Widget>[
           new Row(
             children: <Widget>[
-              new RaisedButton(
+              new ElevatedButton(
+                style: ButtonStyle(
+                  backgroundColor: MaterialStateProperty.all(Colors.blue),
+                ),
                 onPressed: () {
                   _controller.previous();
                 },
-                color: Colors.blue,
+                //color: Colors.blue,
                 child: new Text("Preious"),
               ),
               new SizedBox(
                 width: 8.0,
               ),
-              new RaisedButton(
+              new ElevatedButton(
+                style: ButtonStyle(
+                  backgroundColor: MaterialStateProperty.all(Colors.blue),
+                ),
                 onPressed: () {
                   _controller.next();
                 },
-                color: Colors.blue,
+                //color: Colors.blue,
                 child: new Text("Next"),
               ),
               new SizedBox(
                 width: 8.0,
               ),
-              new RaisedButton(
+              new ElevatedButton(
+                style: ButtonStyle(
+                  backgroundColor: MaterialStateProperty.all(Colors.blue),
+                ),
                 onPressed: () {
                   showModalBottomSheet(
                       context: context,
@@ -135,7 +144,7 @@ class _MyHomePageState extends State<MyHomePage> {
                             children: _types.map((t) => new Text(t)).toList());
                       });
                 },
-                color: Colors.blue,
+                //color: Colors.blue,
                 child: new Text("Change Animation"),
               ),
             ],
