@@ -8,14 +8,14 @@ class IndexController extends ChangeNotifier {
   static const int PREVIOUS = -1;
   static const int MOVE = 0;
 
-  Completer _completer;
+  late Completer _completer;
 
-  int index;
-  bool animation;
-  int event;
+  late int index;
+  late bool animation;
+  late int event;
 
   Future move(int index, {bool animation: true}) {
-    this.animation = animation ?? true;
+    this.animation = animation;
     this.index = index;
     this.event = MOVE;
     _completer = new Completer();
@@ -25,7 +25,7 @@ class IndexController extends ChangeNotifier {
 
   Future next({bool animation: true}) {
     this.event = NEXT;
-    this.animation = animation ?? true;
+    this.animation = animation;
     _completer = new Completer();
     notifyListeners();
     return _completer.future;
@@ -33,7 +33,7 @@ class IndexController extends ChangeNotifier {
 
   Future previous({bool animation: true}) {
     this.event = PREVIOUS;
-    this.animation = animation ?? true;
+    this.animation = animation;
     _completer = new Completer();
     notifyListeners();
     return _completer.future;
