@@ -88,9 +88,9 @@ class ParallaxColor extends StatefulWidget {
   final TransformInfo info;
 
   ParallaxColor({
-    @required this.colors,
-    @required this.info,
-    @required this.child,
+    required this.colors,
+    required this.info,
+    required this.child,
   });
 
   @override
@@ -106,12 +106,12 @@ class ParallaxContainer extends StatelessWidget {
   final double opacityFactor;
 
   ParallaxContainer(
-      {@required this.child,
-      @required this.position,
+      {required this.child,
+      required this.position,
       this.translationFactor: 100.0,
-      this.opacityFactor: 1.0})
-      : assert(position != null),
-        assert(translationFactor != null);
+      this.opacityFactor: 1.0});
+  // : assert(position != null),
+  //   assert(translationFactor != null);
 
   @override
   Widget build(BuildContext context) {
@@ -129,9 +129,11 @@ class ParallaxImage extends StatelessWidget {
   final Image image;
   final double imageFactor;
 
-  ParallaxImage.asset(String name, {double position, this.imageFactor: 0.3})
-      : assert(imageFactor != null),
-        image = Image.asset(name,
+  ParallaxImage.asset(
+    String name, {
+    required double position,
+    this.imageFactor: 0.3,
+  }) : image = Image.asset(name,
             fit: BoxFit.cover,
             alignment: FractionalOffset(
               0.5 + position * imageFactor,
