@@ -58,17 +58,18 @@ class TransformInfo {
   /// Copy from [TransformerPageView.scrollDirection]
   final Axis scrollDirection;
 
-  TransformInfo(
-      {this.index,
-      this.position,
-      this.width,
-      this.height,
-      this.activeIndex,
-      this.fromIndex,
-      this.forward,
-      this.done,
-      this.viewportFraction,
-      this.scrollDirection});
+  TransformInfo({
+    required this.index,
+    required this.position,
+    required this.width,
+    required this.height,
+    required this.activeIndex,
+    required this.fromIndex,
+    required this.forward,
+    required this.done,
+    required this.viewportFraction,
+    required this.scrollDirection,
+  });
 }
 
 abstract class PageTransformer {
@@ -86,9 +87,8 @@ typedef Widget PageTransformerBuilderCallback(Widget child, TransformInfo info);
 class PageTransformerBuilder extends PageTransformer {
   final PageTransformerBuilderCallback builder;
 
-  PageTransformerBuilder({bool reverse: false, @required this.builder})
-      : assert(builder != null),
-        super(reverse: reverse);
+  PageTransformerBuilder({bool reverse: false, required this.builder})
+      : super(reverse: reverse);
 
   @override
   Widget transform(Widget child, TransformInfo info) {
@@ -106,15 +106,15 @@ class TransformerPageController extends PageController {
     bool keepPage = true,
     double viewportFraction = 1.0,
     this.loop: false,
-    this.itemCount,
+    required this.itemCount,
     this.reverse: false,
   }) : super(
             initialPage: TransformerPageController._getRealIndexFromRenderIndex(
-                initialPage ?? 0, loop, itemCount, reverse),
+                initialPage, loop, itemCount, reverse),
             keepPage: keepPage,
             viewportFraction: viewportFraction);
 
-  int getRenderIndexFromRealIndex(num index) {
+  int getRenderIndexFromRealIndex(int index) {
     return _getRenderIndexFromRealIndex(index, loop, itemCount, reverse);
   }
 
@@ -124,7 +124,7 @@ class TransformerPageController extends PageController {
   }
 
   static _getRenderIndexFromRealIndex(
-      num index, bool loop, int itemCount, bool reverse) {
+      int index, bool loop, int itemCount, bool reverse) {
     if (itemCount == 0) return 0;
     int renderIndex;
     if (loop) {
@@ -145,11 +145,11 @@ class TransformerPageController extends PageController {
 
   double get realPage {
     double page;
-    if (position.maxScrollExtent == null || position.minScrollExtent == null) {
-      page = 0.0;
-    } else {
-      page = super.page;
-    }
+    // if (position.maxScrollExtent == null || position.minScrollExtent == null) {
+    //   page = 0.0;
+    // } else {
+    page = super.page ?? 0.0;
+    //}
 
     return page;
   }
@@ -179,12 +179,12 @@ class TransformerPageController extends PageController {
         : realPage;
   }
 
-  int getRealIndexFromRenderIndex(num index) {
+  int getRealIndexFromRenderIndex(int index) {
     return _getRealIndexFromRenderIndex(index, loop, itemCount, reverse);
   }
 
   static int _getRealIndexFromRenderIndex(
-      num index, bool loop, int itemCount, bool reverse) {
+      int index, bool loop, int itemCount, bool reverse) {
     int result = reverse ? (itemCount - index - 1) : index;
     if (loop) {
       result += kMiddleValue;
@@ -197,7 +197,7 @@ class TransformerPageView extends StatefulWidget {
   /// Create a `transformed` widget base on the widget that has been passed to  the [PageTransformer.transform].
   /// See [TransformInfo]
   ///
-  final PageTransformer transformer;
+  final PageTransformer? transformer;
 
   /// Same as [PageView.scrollDirection]
   ///
@@ -205,7 +205,7 @@ class TransformerPageView extends StatefulWidget {
   final Axis scrollDirection;
 
   /// Same as [PageView.physics]
-  final ScrollPhysics physics;
+  final ScrollPhysics? physics;
 
   /// Set to false to disable page snapping, useful for custom scroll behavior.
   /// Same as [PageView.pageSnapping]
@@ -213,12 +213,12 @@ class TransformerPageView extends StatefulWidget {
 
   /// Called whenever the page in the center of the viewport changes.
   /// Same as [PageView.onPageChanged]
-  final ValueChanged<int> onPageChanged;
+  final ValueChanged<int>? onPageChanged;
 
-  final IndexedWidgetBuilder itemBuilder;
+  final IndexedWidgetBuilder? itemBuilder;
 
   // See [IndexController.mode],[IndexController.next],[IndexController.previous]
-  final IndexController controller;
+  final IndexController? controller;
 
   /// Animation duration
   final Duration duration;
@@ -226,7 +226,7 @@ class TransformerPageView extends StatefulWidget {
   /// Animation curve
   final Curve curve;
 
-  final TransformerPageController pageController;
+  final TransformerPageController? pageController;
 
   /// Set true to open infinity loop mode.
   final bool loop;
@@ -253,9 +253,9 @@ class TransformerPageView extends StatefulWidget {
   /// [itemBuilder] will be called only with indices greater than or equal to
   /// zero and less than [itemCount].
   TransformerPageView({
-    Key key,
-    this.index,
-    Duration duration,
+    Key? key,
+    this.index = 0,
+    Duration? duration,
     this.curve: Curves.ease,
     this.viewportFraction: 1.0,
     this.loop: false,
@@ -267,29 +267,26 @@ class TransformerPageView extends StatefulWidget {
     this.transformer,
     this.itemBuilder,
     this.pageController,
-    @required this.itemCount,
-  })  : assert(itemCount != null),
-        assert(itemCount == 0 || itemBuilder != null || transformer != null),
-        this.duration =
+    required this.itemCount,
+  })   : this.duration =
             duration ?? new Duration(milliseconds: kDefaultTransactionDuration),
         super(key: key);
 
   factory TransformerPageView.children(
-      {Key key,
-      int index,
-      Duration duration,
+      {required Key key,
+      required int index,
+      required Duration duration,
       Curve curve: Curves.ease,
       double viewportFraction: 1.0,
       bool loop: false,
       Axis scrollDirection = Axis.horizontal,
-      ScrollPhysics physics,
+      required ScrollPhysics physics,
       bool pageSnapping = true,
-      ValueChanged<int> onPageChanged,
-      IndexController controller,
-      PageTransformer transformer,
-      @required List<Widget> children,
-      TransformerPageController pageController}) {
-    assert(children != null);
+      required ValueChanged<int> onPageChanged,
+      required IndexController controller,
+      PageTransformer? transformer,
+      required List<Widget> children,
+      TransformerPageController? pageController}) {
     return new TransformerPageView(
       itemCount: children.length,
       itemBuilder: (BuildContext context, int index) {
@@ -315,8 +312,12 @@ class TransformerPageView extends StatefulWidget {
     return new _TransformerPageViewState();
   }
 
-  static int getRealIndexFromRenderIndex(
-      {bool reverse, int index, int itemCount, bool loop}) {
+  static int getRealIndexFromRenderIndex({
+    required bool reverse,
+    required int index,
+    required int itemCount,
+    required bool loop,
+  }) {
     int initPage = reverse ? (itemCount - index - 1) : index;
     if (loop) {
       initPage += kMiddleValue;
@@ -324,12 +325,13 @@ class TransformerPageView extends StatefulWidget {
     return initPage;
   }
 
-  static PageController createPageController(
-      {bool reverse,
-      int index,
-      int itemCount,
-      bool loop,
-      double viewportFraction}) {
+  static PageController createPageController({
+    required bool reverse,
+    required int index,
+    required int itemCount,
+    required bool loop,
+    required double viewportFraction,
+  }) {
     return new PageController(
         initialPage: getRealIndexFromRenderIndex(
             reverse: reverse, index: index, itemCount: itemCount, loop: loop),
@@ -338,45 +340,45 @@ class TransformerPageView extends StatefulWidget {
 }
 
 class _TransformerPageViewState extends State<TransformerPageView> {
-  Size _size;
-  int _activeIndex;
-  double _currentPixels;
+  Size? _size;
+  late int _activeIndex;
+  late double _currentPixels;
   bool _done = false;
 
   ///This value will not change until user end drag.
-  int _fromIndex;
+  late int _fromIndex;
 
-  PageTransformer _transformer;
+  PageTransformer? _transformer;
 
-  TransformerPageController _pageController;
+  late TransformerPageController _pageController;
 
   Widget _buildItemNormal(BuildContext context, int index) {
     int renderIndex = _pageController.getRenderIndexFromRealIndex(index);
-    Widget child = widget.itemBuilder(context, renderIndex);
+    Widget child = widget.itemBuilder!(context, renderIndex);
     return child;
   }
 
   Widget _buildItem(BuildContext context, int index) {
     return new AnimatedBuilder(
         animation: _pageController,
-        builder: (BuildContext c, Widget w) {
+        builder: (BuildContext c, Widget? w) {
           int renderIndex = _pageController.getRenderIndexFromRealIndex(index);
-          Widget child;
+          Widget? child;
           if (widget.itemBuilder != null) {
-            child = widget.itemBuilder(context, renderIndex);
+            child = widget.itemBuilder!(context, renderIndex);
           }
           if (child == null) {
             child = new Container();
           }
           if (_size == null) {
-            return child ?? new Container();
+            return child;
           }
 
           double position;
 
           double page = _pageController.realPage;
 
-          if (_transformer.reverse) {
+          if (_transformer?.reverse ?? false) {
             position = page - index;
           } else {
             position = index - page;
@@ -385,8 +387,8 @@ class _TransformerPageViewState extends State<TransformerPageView> {
 
           TransformInfo info = new TransformInfo(
               index: renderIndex,
-              width: _size.width,
-              height: _size.height,
+              width: _size?.width ?? 0,
+              height: _size?.height ?? 0,
               position: position.clamp(-1.0, 1.0),
               activeIndex:
                   _pageController.getRenderIndexFromRealIndex(_activeIndex),
@@ -395,7 +397,7 @@ class _TransformerPageViewState extends State<TransformerPageView> {
               done: _done,
               scrollDirection: widget.scrollDirection,
               viewportFraction: widget.viewportFraction);
-          return _transformer.transform(child, info);
+          return _transformer!.transform(child, info);
         });
   }
 
@@ -446,56 +448,59 @@ class _TransformerPageViewState extends State<TransformerPageView> {
   void _onIndexChanged(int index) {
     _activeIndex = index;
     if (widget.onPageChanged != null) {
-      widget.onPageChanged(_pageController.getRenderIndexFromRealIndex(index));
+      widget.onPageChanged!(_pageController.getRenderIndexFromRealIndex(index));
     }
   }
 
   void _onGetSize(_) {
-    Size size;
-    if (context == null) {
-      onGetSize(size);
-      return;
-    }
-    RenderObject renderObject = context.findRenderObject();
+    Size? size;
+    // if (context == null) {
+    //   onGetSize(size);
+    //   return;
+    // }
+    RenderObject? renderObject = context.findRenderObject();
     if (renderObject != null) {
       Rect bounds = renderObject.paintBounds;
-      if (bounds != null) {
-        size = bounds.size;
-      }
+      //if (bounds != null) {
+      size = bounds.size;
+      //}
+    } else {
+      onGetSize(size);
+      return;
     }
     _calcCurrentPixels();
     onGetSize(size);
   }
 
-  void onGetSize(Size size) {
-    if(mounted){
+  void onGetSize(Size? size) {
+    if (mounted) {
       setState(() {
         _size = size;
       });
     }
-
   }
 
   @override
   void initState() {
     _transformer = widget.transformer;
     //  int index = widget.index ?? 0;
-    _pageController = widget.pageController;
-    if (_pageController == null) {
+
+    if (widget.pageController == null) {
       _pageController = new TransformerPageController(
           initialPage: widget.index,
           itemCount: widget.itemCount,
           loop: widget.loop,
-          reverse:
-              widget.transformer == null ? false : widget.transformer.reverse);
-    }
+          reverse: widget.transformer?.reverse ?? false);
+    } else
+      _pageController = widget.pageController!;
+
     // int initPage = _getRealIndexFromRenderIndex(index);
     // _pageController = new PageController(initialPage: initPage,viewportFraction: widget.viewportFraction);
     _fromIndex = _activeIndex = _pageController.initialPage;
 
     _controller = getNotifier();
     if (_controller != null) {
-      _controller.addListener(onChangeNotifier);
+      _controller?.addListener(onChangeNotifier);
     }
     super.initState();
   }
@@ -503,20 +508,18 @@ class _TransformerPageViewState extends State<TransformerPageView> {
   @override
   void didUpdateWidget(TransformerPageView oldWidget) {
     _transformer = widget.transformer;
-    int index = widget.index ?? 0;
+    int index = widget.index;
     bool created = false;
     if (_pageController != widget.pageController) {
       if (widget.pageController != null) {
-        _pageController = widget.pageController;
+        _pageController = widget.pageController!;
       } else {
         created = true;
         _pageController = new TransformerPageController(
             initialPage: widget.index,
             itemCount: widget.itemCount,
             loop: widget.loop,
-            reverse: widget.transformer == null
-                ? false
-                : widget.transformer.reverse);
+            reverse: widget.transformer?.reverse ?? false);
       }
     }
 
@@ -529,15 +532,15 @@ class _TransformerPageViewState extends State<TransformerPageView> {
       }
     }
     if (_transformer != null)
-      WidgetsBinding.instance.addPostFrameCallback(_onGetSize);
+      WidgetsBinding.instance?.addPostFrameCallback(_onGetSize);
 
     if (_controller != getNotifier()) {
       if (_controller != null) {
-        _controller.removeListener(onChangeNotifier);
+        _controller?.removeListener(onChangeNotifier);
       }
       _controller = getNotifier();
       if (_controller != null) {
-        _controller.addListener(onChangeNotifier);
+        _controller?.addListener(onChangeNotifier);
       }
     }
     super.didUpdateWidget(oldWidget);
@@ -546,11 +549,11 @@ class _TransformerPageViewState extends State<TransformerPageView> {
   @override
   void didChangeDependencies() {
     if (_transformer != null)
-      WidgetsBinding.instance.addPostFrameCallback(_onGetSize);
+      WidgetsBinding.instance?.addPostFrameCallback(_onGetSize);
     super.didChangeDependencies();
   }
 
-  ChangeNotifier getNotifier() {
+  ChangeNotifier? getNotifier() {
     return widget.controller;
   }
 
@@ -582,13 +585,13 @@ class _TransformerPageViewState extends State<TransformerPageView> {
   }
 
   void onChangeNotifier() {
-    int event = widget.controller.event;
+    int event = widget.controller!.event;
     int index;
     switch (event) {
       case IndexController.MOVE:
         {
           index = _pageController
-              .getRealIndexFromRenderIndex(widget.controller.index);
+              .getRealIndexFromRenderIndex(widget.controller!.index);
         }
         break;
       case IndexController.PREVIOUS:
@@ -601,23 +604,22 @@ class _TransformerPageViewState extends State<TransformerPageView> {
         //ignore this event
         return;
     }
-    if (widget.controller.animation) {
+    if (widget.controller!.animation) {
       _pageController
-          .animateToPage(index,
-              duration: widget.duration, curve: widget.curve ?? Curves.ease)
-          .whenComplete(widget.controller.complete);
+          .animateToPage(index, duration: widget.duration, curve: widget.curve)
+          .whenComplete(widget.controller!.complete);
     } else {
       _pageController.jumpToPage(index);
-      widget.controller.complete();
+      widget.controller?.complete();
     }
   }
 
-  ChangeNotifier _controller;
+  ChangeNotifier? _controller;
 
   void dispose() {
     super.dispose();
     if (_controller != null) {
-      _controller.removeListener(onChangeNotifier);
+      _controller?.removeListener(onChangeNotifier);
     }
   }
 }
